+++ viashpy/testing.py
@@ -30,26 +30,50 @@ def meta(test_module):
 
 
 @pytest.fixture
-def executable(meta, test_module):
-    try:
-        return meta["executable"]
-    except KeyError as e:
-        raise KeyError(
-            f"Could not find 'executable' key in 'meta' variable of test module {test_module}. Please make sure it is defined."
-        ) from e
+def executable(meta_attribute_getter):
+    return meta_attribute_getter("executable")
 
 
 @pytest.fixture
-def meta_config_path(meta, test_module):
+def cpus(meta_attribute_getter):
     try:
-        config_path = meta["config"]
-    except KeyError as e:
+        return meta_attribute_getter("cpus")
+    except KeyError:
+        return None
+
+
+@pytest.fixture
+def memory_gb(meta_attribute_getter):
+    try:
+        return meta_attribute_getter("memory_gb")
+    except KeyError:
+        return None
+
+
+@pytest.fixture
+def meta_attribute_getter(meta, test_module):
+    def get_meta_attribute(attr):
+        try:
+            return meta[attr]
+        except KeyError as e:
+            raise KeyError(
+                f"Could not find '{attr}' key in 'meta' variable of test module {test_module}. "
+                "Please make sure it is defined."
+            ) from e
+    return get_meta_attribute
+
+
+@pytest.fixture
+def meta_config_path(meta_attribute_getter, test_module):
+    try:
+        config_path = meta_attribute_getter("config")
+    except KeyError:
         raise KeyError(
             f"The 'config' value was not set in the 'meta' dictionairy of the test module {test_module}."
             "Please define it between the '### VIASH_START ... ### VIASH_END' block."
             "In case this error is reported while using 'viash test' or 'viash_test', "
             "use a viash version >= 0.6.4."
-        ) from e
+        )
     return Path(config_path)
 
 
@@ -81,7 +105,7 @@ def viash_source_config(viash_source_config_path):
 
 
 @pytest.fixture
-def run_component(caplog, executable, viash_source_config_path, viash_executable):
+def run_component(caplog, executable, viash_source_config_path, viash_executable, cpus, memory_gb):
     """
     Returns a function that allows the user to run a viash component.
     The function will use 'viash run' to execute the component or run
@@ -120,7 +144,7 @@ def run_component(caplog, executable, viash_source_config_path, viash_executable
         @run_and_handle_errors
         def wrapper(args_as_list):
             return viash_run(
-                viash_source_config_path, args_as_list, viash_location=viash_executable
+                viash_source_config_path, args_as_list, viash_location=viash_executable, cpus=cpus, memory_gb=memory_gb
             )
 
         return wrapper
@@ -132,6 +156,6 @@ def run_component(caplog, executable, viash_source_config_path, viash_executable
 
     @run_and_handle_errors
     def wrapper(args_as_list):
-        return run_build_component(executable, args_as_list)
+        return run_build_component(executable, args_as_list, cpus=cpus, memory_gb=memory_gb)
 
     return wrapper